requirements.txt
mistralai
python-dotenv
requests
pybase64
streamlit
pandas
numpy
scikit-learn